dice-leap-poc/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dice-leap-poc"
version = "0.1.0"
description = "DICE decision instances → QUBO → local classical solve (Phase 1)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "dimod>=0.12.0",
  "dwave-neal>=0.6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["dice_leap_poc*"]
